fix(bonde-admin-canary): move FormGraphQL to components root
