+++ packages/bonde-admin-canary/src/components/Form/index.js
@@ -1,20 +1,3 @@
-import React from 'react'
-import { reduxForm } from 'redux-form'
-import { InputHint, Flexbox2 as Flexbox } from 'bonde-styleguide'
-
-const Form = ({ children, error, handleSubmit }) => {
-  
-  return (
-    <form style={{ width: '100%' }} onSubmit={handleSubmit}>
-      <Flexbox horizontal end>
-        {error && <InputHint invalid>{error}</InputHint>}
-      </Flexbox>
-      {children}
-    </form>
-  )
-
-}
-
+export { default as Form } from './Form'
+export { default as FormGraphQL } from './FormGraphQL'
 export { Field, SubmissionError } from 'redux-form'
-
-export default reduxForm({ form: 'authLogin' })(Form)
