--- packages/bonde-admin-canary/src/scenes/Auth/scenes/components/index.js
@@ -1,1 +1,0 @@
-export { default as FormGraphQL } from './FormGraphQL'
